Setup to adjust grammar
Grammar can be adjusted and tested interactively
with AntlrWorks. Then CssParserPostBuild will
generate the C# files. Note that to test with Java,
it is necessary to comment out the language=CSharp
in the header and the @members declaration. (Also
the path may need to be adjusted in AntlrWorks for
your machine.)