TD-4299 Avoid having two copies checked into the repo
so future developers don't change one copy and can't figure
out why the changes have no effect. The English is in the
PsSupport folder since it is handled differently on Windows
and Linux. All other langauges are in the
PsSupport\localizations folder as they are handled the same
on Windows and Linux.