Switch to a modern Linux package format
The Debian 1.0 format is deprecated. Also, since the package is maintained as
part of the project, it's appropriate to use the Native format rather than the
Quilt format. The latter uses a diff on top of an original tarball and is
intended for when the packaging is done separately from the original project.