TD-4806 Extra material in reference list
Handle extra material when it occurs in a reference list.
With or without a following book name
The book name may have a digit and space at the beginning
The Extra material can also have a space.
(If the extra material has a space and the book name
is present, this fix will not work.)